Move swagger to pkg
