--- internal/swagger/build.go
@@ -1,3 +1,0 @@
-package swagger
-
-//go:generate /bin/bash "../../hack/autorest.sh"
