+++ pkg/dataplane/swagger/build.go
@@ -1,0 +1,3 @@
+package swagger
+
+//go:generate /bin/bash "../../../hack/autorest.sh"
